User Email Verification, Gallery, Gallery CRUD in ADMIN

+++ WanderLust.Models/DataModels/HappyCustomer.cs
@@ -15,10 +15,10 @@ namespace WanderLust.Models.DataModels
         public string Title { get; set; }
         public string SubTitle { get; set; }
         public string Description { get; set; }
-        public int Flights { get; set; }
-        public int Hotels { get; set; }
-        public int Cars { get; set; }
-        public int Cruises { get; set; }
+        public byte Flights { get; set; }
+        public byte Hotels { get; set; }
+        public byte Cars { get; set; }
+        public byte Cruises { get; set; }
         public virtual AspNetUser AspNetUser { get; set; }
 
     }
